### Does this PR have any customer impact?
no
### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
dev improvements
### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
no
### Are you introducing any new config? If yes, do you have tests with and without them being set?
no
### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description
enable the Rust check-in gh workflow (i've validated the gh action in my personal repo, see https://github.com/AndrewKhoma/documentdb/pull/1)
unify Rust experience across all pipelines
place a rust toolchain version in a single file and reduce the duplication around Rust related components

+++ pg_documentdb_gw/src/auth.rs
@@ -578,5 +578,5 @@ pub async fn get_user_oid(connection_context: &ConnectionContext, username: &str
         .ok_or(DocumentDBError::pg_response_empty())?
         .try_get::<_, tokio_postgres::types::Oid>(0)?;
 
-    Ok(user_oid as u32)
+    Ok(user_oid)
 }
